# Conflicts:
#	composer.lock
#	resources/views/filament/resources/response-resource/pages/browse-entry.blade.php
#	resources/views/filament/resources/response-resource/pages/show-entry.blade.php
#	src/Filament/Resources/FormResource.php
#	src/Filament/Resources/FormResource/Pages/ListForms.php
#	src/Filament/Resources/FormResource/Pages/ViewForm.php
#	src/Models/Category.php
#	src/Models/Field.php
#	src/Models/FieldResponse.php
#	src/Models/Form.php
#	src/Models/Response.php
#	src/Models/Section.php

+++ src/Models/Form.php
@@ -2,8 +2,9 @@
 
 namespace LaraZeus\Bolt\Models;
 
+use Filament\Notifications\Notification;
 use Illuminate\Database\Eloquent\Casts\Attribute;
-use Illuminate\Database\Eloquent\Concerns\HasUlids;
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -40,7 +41,6 @@ class Form extends Model
     use HasActive;
     use HasFactory;
     use HasTranslations;
-    use HasUlids;
     use HasUpdates;
     use SoftDeletes;
 
@@ -59,41 +59,46 @@ class Form extends Model
         'user_id' => 'integer',
     ];
 
-    public function getTable()
+    public function getTable(): string
     {
-        return config('zeus-bolt.table-prefix').'forms';
+        return config('zeus-bolt.table-prefix') . 'forms';
     }
 
     protected static function booted(): void
     {
         static::deleting(function (Form $form) {
+            $canDelete = Extensions::init($form, 'canDelete', []);
+
+            if ($canDelete === null) {
+                $canDelete = true;
+            }
+
+            if (! $canDelete) {
+                Notification::make()
+                    ->title(__('Can\'t delete a form linked to an Extensions'))
+                    ->danger()
+                    ->send();
+
+                return false;
+            }
+
             if ($form->isForceDeleting()) {
-                $form->fieldsResponses()->withTrashed()->get()->each(function ($item) {
-                    $item->forceDelete();
-                });
-                $form->responses()->withTrashed()->get()->each(function ($item) {
-                    $item->forceDelete();
-                });
+                $form->fieldsResponses()->withTrashed()->get()->each(fn ($item) => $item->forceDelete());
+                $form->responses()->withTrashed()->get()->each(fn ($item) => $item->forceDelete());
                 $form->sections()->withTrashed()->get()->each(function ($item) {
-                    $item->fields()->withTrashed()->get()->each(function ($item) {
-                        $item->forceDelete();
-                    });
+                    $item->fields()->withTrashed()->get()->each(fn ($item) => $item->forceDelete());
                     $item->forceDelete();
                 });
             } else {
-                $form->fieldsResponses->each(function ($item) {
-                    $item->delete();
-                });
-                $form->responses->each(function ($item) {
-                    $item->delete();
-                });
+                $form->fieldsResponses->each(fn ($item) => $item->delete());
+                $form->responses->each(fn ($item) => $item->delete());
                 $form->sections->each(function ($item) {
-                    $item->fields->each(function ($item) {
-                        $item->delete();
-                    });
+                    $item->fields->each(fn ($item) => $item->delete());
                     $item->delete();
                 });
             }
+
+            return true;
         });
     }
 
@@ -102,7 +107,7 @@ class Form extends Model
         return 'slug';
     }
 
-    protected static function newFactory(): FormFactory
+    protected static function newFactory(): Factory
     {
         return FormFactory::new();
     }
@@ -185,7 +190,7 @@ class Form extends Model
         );
     }
 
-    public function getUrl(): string|array
+    public function getUrl(): string | array
     {
         if ($this->extensions === null) {
             return route('bolt.form.show', ['slug' => $this->slug]);
